lessons screen ui changed to material
