+++ app/src/main/java/com/nailorsh/repeton/currentlesson/presentation/ui/components/LessonSubject.kt
@@ -1,0 +1,55 @@
+package com.nailorsh.repeton.currentlesson.presentation.ui.components
+
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.res.dimensionResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.nailorsh.repeton.R
+import com.nailorsh.repeton.core.ui.theme.SubjectColor
+
+@Composable
+fun LessonSubject(subject: String, modifier: Modifier = Modifier) {
+    Box(
+        modifier
+            .border(
+                width = 1.dp,
+                shape = MaterialTheme.shapes.small,
+                color = MaterialTheme.colorScheme.onBackground
+            )
+            .background(
+                color = MaterialTheme.colorScheme.primaryContainer,
+                shape = RoundedCornerShape(8.dp)
+            )
+            .padding(horizontal = dimensionResource(R.dimen.padding_medium), vertical = dimensionResource(R.dimen.padding_small))
+
+
+
+
+
+
+
+
+
+    ) {
+        Text(
+            text = subject,
+            style = MaterialTheme.typography.headlineSmall,
+            color = MaterialTheme.colorScheme.onPrimaryContainer,
+            fontWeight = FontWeight.SemiBold,
+            letterSpacing = 1.sp,
+        )
+    }
+}
